pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Gendiff (GENerator of DIFFerences) - a program defining the difference between two data structures (JSON or YAML) and generating new structure containing differences details (including unchanged)."
authors = ["Denis <Dysprosium92@ya.ru>"]
license = "MIT"
packages = [
    { include = "gendiff" },
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.main:main"

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.4.1"
wemake-python-styleguide = "^0.15.2"
pyformat = "^0.7"
pytest-cov = "^2.12.0"
pytest = "^6.2.4"
coverage = "^5.5"
rope = "^0.19.0"
black = "^21.5b1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
